pyproject.toml
[project]
name = "knowledgebase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diskcache>=5.6.3",
    "fastapi>=0.120.1",
    "langchain-astradb>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "pdfplumber>=0.11.7",
    "chromadb>=0.5.18",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "volcengine-python-sdk[ark]>=4.0.29",
    "tqdm>=4.67.1",
    "spacy>=3.8.7",
    "pip>=25.3",
    "loguru>=0.7.3",
    "python-multipart>=0.0.20",
]
